# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Synthora API",
    "description": "Multi-model AI orchestration — answers that surpass any single model",
    "version": "2.8.8"
  },
  "paths": {
...
    "/api/admin/usage": {
      "get": {
        "summary": "Admin Usage",
        "description": "Get usage stats for all users. Admin only.",
        "operationId": "admin_usage_api_admin_usage_get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUsageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/usage/{user_id}": {
      "get": {
        "summary": "Admin User Usage",
        "description": "Get usage history for a specific user. Admin only.",
        "operationId": "admin_user_usage_api_admin_usage__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 7,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserUsageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/cost-report": {
      "get": {
        "summary": "Admin Cost Report",
        "description": "Actual API cost report for today (or given date). Admin only.\n\nv4.30: Cost figures are now read from per-query estimated_cost_usd stored in\nquery_history (actual tracked cost, not P50 estimates). P50 fallback is used\nonly when no history records exist for the date (e.g. new deployment).\n\nReturns per-mode call counts, actual API cost, and top users by consumption.",
        "operationId": "admin_cost_report_api_admin_cost_report_get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCostReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminCostReportResponse": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "mode_call_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Mode Call Counts"
          },
          "mode_costs_usd": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Mode Costs Usd"
          },
          "total_cost_usd": {
            "type": "number",
            "title": "Total Cost Usd"
          },
          "top_users_by_queries": {
            "items": {
              "$ref": "#/components/schemas/AdminTopUserQueriesItem"
            },
            "type": "array",
            "title": "Top Users By Queries"
          },
          "top_users_by_cost": {
            "items": {
              "$ref": "#/components/schemas/AdminTopUserCostItem"
            },
            "type": "array",
            "title": "Top Users By Cost"
          },
          "cost_basis": {
            "type": "string",
            "title": "Cost Basis"
          }
        },
        "type": "object",
        "required": [
          "date",
          "mode_call_counts",
          "mode_costs_usd",
          "total_cost_usd",
          "top_users_by_queries",
          "top_users_by_cost",
          "cost_basis"
        ],
        "title": "AdminCostReportResponse"
      },
      "AdminTopUserCostItem": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "query_count": {
            "type": "integer",
            "title": "Query Count"
          },
          "total_cost_usd": {
            "type": "number",
            "title": "Total Cost Usd"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "query_count",
          "total_cost_usd"
        ],
        "title": "AdminTopUserCostItem"
      },
      "AdminTopUserQueriesItem": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "total_queries": {
            "type": "integer",
            "title": "Total Queries"
          },
          "light": {
            "type": "integer",
            "title": "Light",
            "default": 0
          },
          "socratic": {
            "type": "integer",
            "title": "Socratic",
            "default": 0
          },
          "deep": {
            "type": "integer",
            "title": "Deep",
            "default": 0
          },
          "research": {
            "type": "integer",
            "title": "Research",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "total_queries"
        ],
        "title": "AdminTopUserQueriesItem"
      },
      "AdminUsageResponse": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "limits": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Limits"
          },
          "users": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "integer"
              },
              "type": "object"
            },
            "type": "object",
            "title": "Users"
          }
        },
        "type": "object",
        "required": [
          "date",
          "limits",
          "users"
        ],
        "title": "AdminUsageResponse"
      },
      "AdminUserUsageResponse": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "today": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Today"
          },
          "limits": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Limits"
          },
          "history": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "integer"
              },
              "type": "object"
            },
            "type": "object",
            "title": "History"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "today",
          "limits",
          "history"
        ],
        "title": "AdminUserUsageResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}